callback: create host once and cache the response
The callback was doing an API call to the host endpoint for every task
result to retrieve the host ID so it could attach the result to the
host.
It later needed to retrieve IDs when attaching statistics as well.

For a simple 3 task playbook against 64 hosts, this brings down the
amount of API calls to /api/v1/hosts from 256 to 64.

See the issue [1] on GitHub for details.

[1]: https://github.com/ansible-community/ara/issues/80

Change-Id: If9c65e224b758a0ee5707dd8270b4ee83ddfa5bd