api: move duration to the database model
Before this change, the duration was not stored in the database and was
instead calculated by the API based on the start and end date of
playbooks.

This change moves the duration to the database so it is no longer
computed by the API. This will allow us to search, filter and sort by
duration.

Change-Id: I34e64da98f05ac3c8d0c5e18436d6b16e62ae787