api: Fix UnicodeEncodeError when parsing and rendering results
This is an attempt to prevent encoding errors when posting results or
when rendering them in the API and built-in interface.

Setting UNICODE_JSON [1] to false allows the callback to post results'
content and allows the API browsing interface to render the result.

The built-in UI had a similar problem but isn't fixed by the
UNICODE_JSON because we are using a Response object so use
"surrogateescape" instead.

Related: https://github.com/ansible-community/ara/issues/48

[1]: https://www.django-rest-framework.org/api-guide/settings/#unicode_json

Change-Id: I48bcef440a1ee9c8574fdd24d5c05ef0f13ca666