ara_api role: add support for installing from fedora packages
This patch is self tested by the addition of a new job that exercises
the role with the packages but it does not test the current source yet.

In order to test the current source, we'll need to set up a job that
builds a package from the current source and then configure the job to
install from these built packages.

Change-Id: Ic326a5243b55b415cd31b8a5702fe92d5dfe0651