Add a job to run playbooks and record them on the demo
This adds a secret for authenticating with the live demo instance,
and a job which runs basic integration tests. The results of these tests
are sent to the live demo instance.

The job is set to run in the "post" Zuul pipeline so the secret cannot be
exposed by speculatively testing a job change.

Change-Id: I8a7926ed547b1acb4b7021ca7c63abb9d68c0ac9