callback: improve consistency of stored task results
CallbackBase._dump_results() is oriented toward screen display and
throws away data based on the verbosity level of ansible-playbook,
whereas I would expect the data stored by ARA to be unaffected by
Ansible's display settings. It also unconditionally throws away
exception information.

The useful things it was doing for us are simple enough to do
directly.

The sanitization of ansible_facts should be done before the result is
stored, not after.

The self.loop_items processing was dead code (nothing in
this callback or CallbackBase added items to this list.)

Change-Id: I110a94b4e0231a0f37aceebb838889bbfa88d661