settings: align with dynaconf preference for ruamel.yaml
When installing from pip with extras, dyanconf[yaml] used to pick up
pyyaml but it recently started preferring ruamel.yaml after the recent
3.0.0 release.

Our usage doesn't make us very opinionated since we only use it for
settings so align with dynaconf.

It was necessary to adjust the fedora-source container script so it
could pick up the fix from the patch itself instead of cloning from
master.

For the fedora-pypi image, it won't be able to use this patch until
it's released so install PyYAML in the meantime.

The distributed sqlite CI job is failing for unrelated reasons but it's
also in the process of being migrated to the new collection so make it
non-voting for now.

Related: https://github.com/ansible-community/ara/issues/146
Change-Id: I83c585c7e946c02255beab1174367e550b3e798b