container-images: incorporate, test and publish centos8
This leverages the contribution from Fabian [1] for building an image
based on CentOS8.
The image will be built and tested for every new patch with Zuul and
it'll be pushed to DockerHub [2] after every merge.

The tag nomenclature had to be changed in order to fit the distribution
as an attribute, before:
- latest (tagged from source-latest)
- pypi-latest
- distribution-latest
- source-latest

to:
- latest (tagged from fedora32-source-latest)
- fedora32-source-latest
- fedora32-distribution-latest
- fedora32-pypi-latest
- centos8-pypi-latest

[1]: https://github.com/ansible-community/ara/commit/8491cc5a72e24de5e1f608f39ba7c0997601c279
[2]: https://hub.docker.com/r/recordsansible/ara-api

Change-Id: I971abd97039a28e825b6b7cd43a7559abe709f30