CLI: Add '--long' argument to ara result list
We have to be opinionated about the columns we return by default because
we do not have unlimited width to display data in.

The 'changed' and 'ignore_errors' fields are somewhat combined into the
status field that returns 'changed' or 'ignored' while the 'play' field
is not necessarily required if the task and host context are present.

Change-Id: I6c26424a53bb89df18a2f4d7aafe42fe93a2b171