CLI: Add API filtering arguments to 'ara playbook prune'
These are mainly ported from 'ara playbook list'.

The API support setting the --limit argument to any number and it will
override pagination settings.

Here we are defaulting the limit to 200 instead of a very high number.
It's still meaningfully more than the default pagination setting of 100
while keeping the scale of the damage down in case of footgun scenarios.

Users are free to provide a higher limit if they want, ex: --limit 9000.

Fixes: https://github.com/ansible-community/ara/issues/166
Change-Id: I0a2aab6fedeea758a64b71452f7e249fb8ac516c