callback: push non-blocking calls to threads
There are some parts of the callback that need to be synchronous because
they rely on an object that must be created before, for example.

However, for the parts that don't need to run synchronously, thread them
so we don't block Ansible from running.

This results in a dramatic performance increase.

Related: https://github.com/ansible-community/ara/issues/171
Change-Id: I4317d1d1175a5286704b04f57de9ef73fa911a3a