API: Remove "changed" and "ignored" statuses
The principal consumers of the API relied on a combination of the
status as well as the fields "changed" and "ignore_errors" to infer a
"changed" or "ignored" status.

This lead users to believe they could search for these statuses and it
wouldn't work. The consumers should expose the status differently but
that will be in another patch.

Change-Id: Ib34a65179d07186ab2d2efc65ce14c46e2596df0