callback: set a more accurate start date for results
Because we always set the started date of results as the beginning of
the task, each host result for the task apparently took a little bit
longer than the previous to complete because they were running one after
the other.

By storing a timestamp at the end of a result, we can use it as the
start date of the new result. While imperfect, it will be far more
accurate than relying on the task started date to calculate the
duration of each result.

Fixes: https://github.com/ansible-community/ara/issues/173
Change-Id: Iade5dc3ba35272ac7993a114d06bbefcd096d677