UI: Link playbooks to filtered results based on status
This will apply a default filter to playbook links based on its status.

When failed, it will search for results with the status "failed" or
"unreachable" -- otherwise, it will search for results that have
resulted in a change.

The objective is to highlight changed, failed and unreachable results in
an ocean of skipped and ok results.

Change-Id: I662682296be68e387a26da16be8313daa15584bf