callback: improve accuracy of result durations with Ansible >=2.8
Ansible >= 2.8 added the "v2_runner_on_start" callback hook which gives
us an accurate result start date instead of the task start date.

When running Ansible 2.7 and below, ara will fall back to the task start
date which was the original behavior.

In addition, recover the ended date for the result slightly sooner for
better accuracy.

Fixes: https://github.com/ansible-community/ara/issues/173
Change-Id: I43295fda19eb900a28df023bf6c4680e62a7f91e