Remove Ansible roles from the repository
These have been moved out to an Ansible collection:
https://github.com/ansible-community/ara-collection

Leaves a README.rst to tell users where to look for for a while.

Because the existing Zuul CI jobs relied on these roles, this also
changes the Zuul config to use the jobs that are used to test the
collection instead.

Related: https://github.com/ansible-community/ara/issues/71
Change-Id: I1c573d8677d36cfd431101f1f181253c1cce3bc1