Re-add support for static generation
This is more or less the same implementation as previously though this
one is made a little bit less redundant with the addition of a
"static_url" filter.

Whereas before we used a mixture of "manual" links (such as
../playbooks/<id>.html), now we'll always use the django 'url' tag and
then adjust the destination with the static_url filter if we are generating
a static report.

It's not very clean due to a lot of conditionals but it works and maybe
we can find a better way later.