ui: Fix truncatepath calculations
`len(dirname) + len(basename)` is 1 less than `len(path)`, so this
could end up returning the entire path that was passed in with "..."
prepended. Plus, to always return something that is less than `count`
we need to account for the prepended string and handle the case where
the filename on its own is very long.

Some of the things passed to this function are not actually paths but
should probably still be shortened.

Change-Id: Ib7e34d8dd70ec62cb319a82e44da4e7f269ae480