docs: Add a database model graph
This is an addition to help flesh out the docs about relationships
between objects and a quick win since it works out of the box with
django_extensions' graph_models.

Since the database model doesn't move very much, it's rendered a single
time manually (instead of, say, automatically after every commit) and
there are instructions included to generate it again, if need be, in order
to avoid adding django_extensions as a dependency.