api/callback/ui: Add support for tracking task delegation
ara did not previously record whether or not a task had been delegated
and if so, to which host. Now it does.

This adds a field for it in the result model so we can display it when
browsing a result's details.

From an API perspective, it also adds support for searching results that
may have been delegated to a particular host ID like so:
    /api/v1/results?delegated_to=<id>