Add caching for tasks in callback plugin (#268)
When running a playbook with free strategy, multiple hosts could be running
different playbook tasks at the same time.

This caused ARA to associate results to the wrong task id, resuting in task
names not matching the task result in the ARA web interface.

Now, the task is cached by it's uuid. The task result includes the task uuid, so
this is used to ensure the correct task id is associate with a particular result.