Add search for all reports and by status, fix search by date
It adds a new checkbox to toggle between the latest reports for each
host (/api/v1/distincthosts) and for every report for every
host (/api/v1/hosts).

Clicking on a host name searches for that host and toggles ?latest=false
in order to display all reports for that host.

The status checkboxes in reality are a check against the value of each
status field, the search is implemented with __gt operators, i.e:
    /hosts?name=localhost&latest=false&changed__gt=0

Would only return hosts named localhost that have more than 0 changed
results.