api: rename distincthosts to latesthosts
This has no incidence on functionality and is just a name change intended
to clarify and streamline the meaning and vocabulary around the feature.

Depending on the context, the words "latest" and "distinct" could have
different meanings or be used interchangeably.

Instead of having the notion of a DistinctHost which has a "latest"
that in reality is a "host", do away with the potential confusion and
name it LatestHost with a "host" reference instead of "latest".