ui: Fix latest host report ordering by date
The intent was to order hosts by their latest update with the most
recent at the top: -updated.

However, it turns out that we were sorting LatestHost objects instead of
Host objects. This commits fixes it so we sort based on the date of the
host, not the latesthost reference.