api: Update LatestHost when deleting a playbook
We have a custom perform_destroy that runs when deleting a host via the
API (DELETE /api/v1/hosts/<id>) that takes care of updating the
latesthost entry.

However, because hosts get deleted via the database model
(on_delete=models.CASCADE) when deleting a playbook via the API, the
latesthost entry would get deleted when the host was deleted instead of
being updated.

This adds a very similar perform_destroy for playbooks that iterates
over the hosts and if one of the hosts happens to be a LatestHost, it
will update the entry before deleting.