callback: skip recording delegate_to for skipped loop items
8df13a99ba41ecac4040db11caa867e2c9c98b72 avoided recording tasks with
delegate_to that had been skipped entirely due to a 'when' condition
because Ansible doesn't template the variable sent to the callback.

It turns out that we also need to avoid recording it for skipped items
while in a loop because the delegate_to also doesn't get templated by
Ansible.