callback/api/cli: Add a "failed" status for tasks
Before this change, ara only recorded whether the task
was completed or not.

ara will now infer the status of a task based on whether there
was one or more failed or unreachable result from any of the hosts
that ran this task.

This approach works with the serial strategy with a known issue
for the free strategy where it might flip the status of a task
as the results roll in out of order.