callback: truncate names if they exceed 255 characters
Documentation suggests (https://www.freesoft.org/CIE/RFC/1035/9.htm)
that DNS hostnames cannot be longer than 255 characters but there are no
such limits for Ansible inventory hostnames.

Trying to send an inventory hostname that is longer than 255 characters
to the API trips into an exception because the field is a varchar(255).

Truncate the hostname to prevent an exception and warn about it once per
host if it happens.

Playbook names, play names and label names have the same restriction to
255 characters. We have tried raising the max length but varchar fields
longer than 255 characters are not supported with the MySQL backend so
truncate those too.