ui: Adjust host headers and move facts after results
The card header font was smaller than elsewhere in the UI.
Elsewhere is h4 and h4 is equivalent to 1.40625rem.

I couldn't find out how to allow the card headers to be responsive
without relying on a table (like in the playbook card) while also being
a button which allows to collapse the card.

This is mostly an issue on mobile and we can improve this later.

Based on nothing more than personal experience, the task results should
be more relevant than host facts which means less scrolling to see what
was the last task that ran on this host immediately.