ui: truncate long controller hostnames
Fields that can be very long are generally truncated in tables
with the full version available on mouse-over.

This is a compromise to avoid a single column taking up most of the
width available in the table.

controller hostnames can be up to 255 characters so truncate them like
we do elsewhere.