maintenance: Bump version of django from 3.2LTS to 4.2LTS
Django 4.2LTS was released in April 2023 and 3.2LTS will be supported
until April 2024: https://docs.djangoproject.com/en/4.2/releases/4.2/

Django 4.2 raises the Python requirement from 3.6 up to 3.8 so in effect
it means we are also raising the requirement for ara itself.

Given the change in requirements, this will warrant a version bump for
ara from 1.6.1 to 1.7.0.