settings: Use ruamel.yaml's non-deprecated version of .dump
We used .dump by itslef which used to be compatible for both ruamel.yaml
and pyyaml but it has since been deprecated in ruamel.yaml>=0.18.

Since we are only using ruamel.yaml (as part of dynaconf) let's stop
trying to be compatible with pyyaml and just use ruamel's way of doing
it.