docs: refresh getting started docs
The getting started guide was originally a copy from the README
and it accidently became stale and out of date as the README was
improved and maintained.

We can make the README even better eventually and there shouldn't
be two sources of truth.

Let's keep the README for this and link to it instead.