callback: use expired status when playbooks are interrupted
The semantic is important: we don't know whether anything has
actually failed but we know the playbook was interrupted.

If anything had failed without 'ignore_errors' or a rescue block, it
would have brought an end to the playbook.

We don't yet have an "interrupted" status for playbooks.

There could be one but for now let's use 'expired' which was created
to actually expire interrupted playbooks that will probably never
finish.