docs: clarify troubleshooting performance with Django authentication
Using the builtin Django authentication system on the API server results
in very significant performance degradation. The documentation already
strongy suggests delegating authentication to a load balancer / reverse
proxy but some people might still rely on the default authentication while
experimenting.

The Troubleshooting section hinted that the performance issue
could be due to Django<->database latency, when it is in fact mostly due
to the use of PBKDF2 for password hashing.