client: Consider that username and password may be empty
We were checking if the variables were set at all.
The variables can be set but to an empty string value.

This can happen when the username value isn't set, for example:
export ARA_API_USERNAME=${USERNAME:""}

If USERNAME isn't set, the value will be an empty string.
This sets the variable, it's just empty.

The client would wind up trying to authenticate with empty usernames and
passwords.