CI: Migrate jobs from GitHub/Zuul to Forgejo actions
This removes the implementation of CI jobs with Zuul and includes
all the plumbing to make our existing jobs run on Codeberg/forgejo:
1) docs, linters, unit and integration tests
2) container image build & tests

The first run inside a Fedora 42 container image.
This image is built with buildah and published to the repository.

The second creates an OpenStack VM (Fedora 42 or Debian 12) and then
runs the builds and tests there.

In addition, where relevant, we are now sending ansible playbook results
to demo.recordsansible.org for reporting.