UI: Redesign playbook header & align hosts, files, records
The playbook header card started out simple when it was
first designed but over time ara recorded more and more information:
the versions, the user, the controller, and so on.

It became cluttered and wasn't very UX friendly on mobile.
We're redesigning it to keep the same amount of information but better
organized and folding down appropriately for mobile view.

Taking an opportunity to align the cards for hosts, files and records
to be consistent with this new header, they will also fold down nicely
on mobile.

In addition:
- lint/autoformat the css file