feature: Bubble up deprecations, exceptions and warnings
If there are any deprecations, exceptions or warnings, these should
bubble up to the task level so they can be individually highlighted.

This contains the entire implementation from the database migration,
the callback update to attach the notes to the task as well as the UI
implementation for tasks, playbook and host results.